fix: macOs behaviour with AppActivate
